requirements.txt
asgiref
# astroid==2.3.3
#ERROR: Cannot install -r requirements.txt (line 5), -r requirements.txt (line 8) and s3transfer==0.2.1 because these package versions have conflicting dependencies.
autopep8
aws-xray-sdk
#awscli==1.16.309
bcrypt
boto
boto3
botocore
certifi
cffi
chardet
colorama
cookies
cryptography
Django
django-s3-storage
docker
docutils
ecdsa
future
geolocation-python
idna
isort
Jinja2
jmespath
jsondiff
jsonpickle
lazy-object-proxy
MarkupSafe
mccabe
mock
moto
Pillow
pyaml
pyasn1
pycodestyle
pycparser
pycryptodome
pylint
pylint-django
pylint-plugin-utils
pypiwin32
python-dateutil
python-jose
pytz
# pywin32==227
PyYAML
requests
responses
rsa
s3transfer
six
sqlparse
timeago
typed-ast
urllib3
websocket-client
Werkzeug
wrapt
xmltodict
